+++ packages/primitives/src/components/Form/index.tsx
@@ -1,11 +1,11 @@
-import Field from '@components/Form/Field';
+import Field from '@components/Form/Field/Field';
 import RadioField from '@components/Form/Field/RadioField';
 import TextField from '@components/Form/Field/TextField';
 import FormRoot from '@components/Form/Form';
 
 const Form = Object.assign(FormRoot, {
-  TextField,
   Field,
+  TextField,
   RadioField,
 });
